requirements.txt
langgraph>=0.0.19
pyjokes
pydantic>=2.0.0
openai>=1.0.0
pyyaml>=6.0